% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_amrs_ddg.R
\name{amrs_ddg}
\alias{amrs_ddg}
\title{Calculate site-dependent ddg profile using amrs method}
\usage{
amrs_ddg(wt, mut_model, mut_dl_sigma, mut_sd_min)
}
\arguments{
\item{wt}{is the (wild-type) protein to mutate (an object obtained using \code{set_enm})}

\item{mut_model}{is the mutational model, which may be "lfenm" or "sclfenm"}

\item{mut_dl_sigma}{is the standard deviation of a normal distribution from which edge-length perturbations are picked (LFENM model).}

\item{mut_sd_min}{is integer sequence-distance cutoff, only edges with \code{sdij >= mut_sd_min} are mutated}
}
\value{
a tibble containing the ddg profile
}
\description{
\code{amrs_ddg()} uses an analytical method to calculate averages over mutations of elastic-network-model stability changes.
}
\details{
If code{ddg[j,m]} is the mutational change in free energy calculated using the lfenm model for a mutation `m` at site `j`,
then this function returns a vector with values \code{ddg[j] = rowMeans(ddg[j,m])}, i.e., averages over mutations.

For details see \doi{10.7717/peerj.11330}
}
\examples{
\dontrun{
pdb <- bio3d::read.pdb("2acy")
wt <- set_enm(pdb, node = "ca", model = "ming_wall", d_max = 10.5, frustrated = FALSE)
responses <- amrs_ddg(wt, mut_model = "lfenm", mut_dl_sigma = 0.3, mut_sd_min = 1, seed = 1024)
}

}
\seealso{
Other mutscan functions:
\code{\link[=admrs]{admrs()}},
\code{\link[=amrs]{amrs()}},
\code{\link[=amrs_ddgact]{amrs_ddgact()}},
\code{\link[=sdmrs]{sdmrs()}},
\code{\link[=smrs]{smrs()}},
\code{\link[=smrs_ddg]{smrs_ddg()}},
\code{\link[=smrs_ddgact]{smrs_ddgact()}}
}
\concept{mutscan functions}
