% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penm-package.R
\docType{package}
\name{penmscan-package}
\alias{penmscan}
\alias{penmscan-package}
\title{penmscan: Build and Perturb Elastic Network Models}
\description{
Functions to calculate ENM models, mutate ENMs by perturbing them, perform single-site mutional scans to calculate average mutation-response matrices, and perform double-site mutational scans to calculate compensation matrices.
}
\details{
The \code{penmscan} package includes functions to calculate various Elastic Network Models for proteins,
    perform normal mode analysis, and using lfenm, obtain mutant proteins and the corresponding mutant ENMs.
    In addition, it has functions to scan the various average-responses w.r.t. single-site mutations and double-site mutations.
}
