% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_sdmrs.R
\name{sdmrs}
\alias{sdmrs}
\title{Calculate a double-mutational-scan matrix numerically (simulation-based)}
\usage{
sdmrs(
  wt,
  nmut,
  mut_dl_sigma,
  mut_sd_min,
  option = "mean_max",
  response = "dr2",
  seed = 1024
)
}
\arguments{
\item{wt}{is the (wild-type) protein to mutate (an object obtained using \code{set_enm})}

\item{nmut}{is the number of mutations per site to simulate}

\item{mut_dl_sigma}{is the standard deviation of a normal distribution from which edge-length perturbations are picked (LFENM model).}

\item{mut_sd_min}{is integer sequence-distance cutoff, only edges with \code{sdij >= mut_sd_min} are mutated}

\item{option}{is either "mean_max" (default) or "max_max", depending on which compensation measure is desired.}

\item{response}{is the response desired, which maybe either "dr2", "de2", or "df2"}

\item{seed}{seed for random generation of mutations}
}
\value{
A compensation matrix, rows are initially mutated site, j is compensation site
}
\description{
Returns a compensation matrix: element (i,j) measures the degree of compensation of structural deformations produced by pairs of mutations at sites i and j.
It uses a simulation method (calculates responses for various instances of forces, then calculates means or maxima)
Two measures are implemented:
"mean_max" (default), the structural compensation maximized over mutations at j and averaged over mutations at i;
"max_max" is the structural compensation maximized over mutations at i and j.
}
\details{
For details see \doi{10.7717/peerj.11330}
}
\examples{
\dontrun{
pdb <- bio3d::read.pdb("2acy")
wt <- set_enm(pdb, node = "ca", model = "ming_wall", d_max = 10.5, frustrated = FALSE)
dmat <- sdmrs(wt, nmut = 10, mut_dl_sigma = 0.3, mut_sd_min = 1, option = "max_max", response = "dr2")
}

}
\seealso{
Other mutscan functions:
\code{\link[=admrs]{admrs()}},
\code{\link[=amrs]{amrs()}},
\code{\link[=amrs_ddg]{amrs_ddg()}},
\code{\link[=amrs_ddgact]{amrs_ddgact()}},
\code{\link[=smrs]{smrs()}},
\code{\link[=smrs_ddg]{smrs_ddg()}},
\code{\link[=smrs_ddgact]{smrs_ddgact()}}
}
\concept{mutscan functions}
