% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_smrs.R
\name{smrs}
\alias{smrs}
\title{Calculate a mutation-response matrix numerically (simulation-based)}
\usage{
smrs(
  wt,
  nmut,
  mut_dl_sigma,
  mut_sd_min,
  option = "site",
  response = "dr2",
  seed = 1024
)
}
\arguments{
\item{wt}{is the (wild-type) protein to mutate (an object obtained using \code{set_enm})}

\item{nmut}{is the number of mutations per site to simulate}

\item{mut_dl_sigma}{is the standard deviation of a normal distribution from which edge-length perturbations are picked (LFENM model).}

\item{mut_sd_min}{is integer sequence-distance cutoff, only edges with \code{sdij >= mut_sd_min} are mutated}

\item{option}{is either "site" (default) or "mode"}

\item{response}{is either "dr2" (default), "de2", or "df2"}

\item{seed}{is a seed for random-number generation of mutations}
}
\value{
A response matrix, columns are mutated sites, rows are responses, of a given site or normal mode.
}
\description{
Returns a mutation-response matrix
It uses a simulation method (averages over perturbations). For details see \doi{10.7717/peerj.11330}
}
\details{
A site-by-site response matrix has elements Mij that measure the response (e.g. deformation) of site i averaged over mutations at site j.
A mode-by-site response matrix has elements Mnj that measure the response (e.g. deformation) along mode n averaged over mutations at site j.


It may calculate either  site-by-site or mode-by site response matrices
Three type of response may be calculated, "dr2" (dr2ij and dr2nj), "de2" (de2ij and de2nj), and "df2" (df2ij and df2nj).
}
\examples{
\dontrun{
pdb <- bio3d::read.pdb("2acy")
wt <- set_enm(pdb, node = "ca", model = "ming_wall", d_max = 10.5, frustrated = FALSE)
mrs_matrix <- smrs(wt, nmut = 10, mut_model = "lfenm", mut_dl_sigma = 0.3, mut_sd_min = 1, response = "dr2", seed = 1024)
}

}
\seealso{
Other mutscan functions:
\code{\link[=admrs]{admrs()}},
\code{\link[=amrs]{amrs()}},
\code{\link[=amrs_ddg]{amrs_ddg()}},
\code{\link[=amrs_ddgact]{amrs_ddgact()}},
\code{\link[=sdmrs]{sdmrs()}},
\code{\link[=smrs_ddg]{smrs_ddg()}},
\code{\link[=smrs_ddgact]{smrs_ddgact()}}
}
\concept{mutscan functions}
