% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutscan_smrs_ddgact.R
\name{smrs_ddgact}
\alias{smrs_ddgact}
\title{Calculate ddgact profile using smrs method}
\usage{
smrs_ddgact(
  wt,
  pdb_site_active,
  nmut,
  mut_model,
  mut_dl_sigma,
  mut_sd_min,
  seed
)
}
\arguments{
\item{wt}{is the (wild-type) protein to mutate (an object obtained using \code{set_enm})}

\item{pdb_site_active}{is a list of sites (using their resno in the pdb file) that define an "active site"}

\item{nmut}{is the number of mutations per site to simulate}

\item{mut_model}{is the mutational model, which may be "lfenm" or "sclfenm"}

\item{mut_dl_sigma}{is the standard deviation of a normal distribution from which edge-length perturbations are picked (LFENM model).}

\item{mut_sd_min}{is integer sequence-distance cutoff, only edges with \code{sdij >= mut_sd_min} are mutated}

\item{seed}{is an integer seed to set_seed before generating random mutations}
}
\value{
a tibble containing the ddg profile
}
\description{
An influence profile is a vector with elements (Vj), where Vj is a total protein response to mutations of site j, average over mutations of j.
It uses a simulation method (calculates responses for various instances of forces, then calculates averages)
`ddgact` is the mutational change in activation free energy calculated using the lfenm model
It is the energy necessary to deform the mutant's active site to "push it back" to its wild-type conformation
}
\details{
For details see \doi{10.7717/peerj.11330}
}
\examples{
\dontrun{
pdb <- bio3d::read.pdb("2acy")
wt <- set_enm(pdb, node = "ca", model = "ming_wall", d_max = 10.5, frustrated = FALSE)
pdb_site_active <- c(10,14,21)
responses <- smrs_ddgact(wt, pdb_site_active, nmut = 10, mut_model = "lfenm", mut_dl_sigma = 0.3, mut_sd_min = 1, seed = 1024)
}

}
\seealso{
Other mutscan functions:
\code{\link[=admrs]{admrs()}},
\code{\link[=amrs]{amrs()}},
\code{\link[=amrs_ddg]{amrs_ddg()}},
\code{\link[=amrs_ddgact]{amrs_ddgact()}},
\code{\link[=sdmrs]{sdmrs()}},
\code{\link[=smrs]{smrs()}},
\code{\link[=smrs_ddg]{smrs_ddg()}}
}
\concept{mutscan functions}
